13 - コンポーネント関連のtestを移動
問題
今のフォルダ構成のままだと、コンポーネントのテストなのかコンポーネント以外のテストなのかわからない。
app.test.jsがコンテナなので、コンポーネントテストが不要

解決策
__components__というフォルダを作り、ファイルを移動する。
app.test.jsを削除する。
