+++ __tests__/__components__/Home.test.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Home from '../src/components/Home';
+import Home from '../../src/components/Home';
 import renderer from 'react-test-renderer';
 
 describe ('Home画面', () => {
